Add binaries to repo
The primary aim of this change is to fix it on the latest version of
macOS: Catalina, which no longer supports 32-bit applications.

In order to fix this, this change bundles the version of `kindlegen`
that is included in Kindle Previewer 3, which has been compiled for
64-bit systems.

While we're here, we also include the binaries for the other systems,
since we trust this repo (because we own it).

We also update the Travis config to run on LTS Node 10, and LTS Linux
Bionic.